Fix linker error when using NT weak externals
Lets take the following example:

a.c:
	void foo(void);

	int main()
	{
		foo();
			return 0;
	}

b.c:
	#include <stdio.h>

	void __attribute__((weak)) foo(void)
	{
		printf("%s weak\n", __func__);
	}

Compiling this with mingw will trigger the following linker error:

$ i686-w64-mingw32-gcc -o a.exe a.c b.c
/tmp/ccZ6Yw7O.o:a.c:(.text+0xc): undefined reference to `foo'
collect2: error: ld returned 1 exit status

Depending on the order in which a.c and b.c are linked the following
is happening in _bfd_generic_link_add_one_symbol:

1. If we link in a.c b.c order then foo will be added in as an
undefined symbol from a.c. Then it will be added as a weak undefined*
symbol from b.c (section undefined, flags BSF_WEAK) and the action
taken will be no action, i.e. keep the symbol undefined.

2. If we link in b.c a.c order then foo will first be added as an
undefined weak symbol from b.c. Then it will be added as undefined
from a.c and the action taken will be to make the symbol undefined.

To fix this issue, this patch checks for an weak external symbol when
doing relocations even if the symbol is undefined. In order to do so,
we also have to also populate the symbol_class and aux members when
adding a weak external symbol over an undefined or weak defined
symbol.

*In PE COFF case defined weak symbols are implemented as weak
externals which requires them to be seen as undefined (weak) symbols.

bfd/

2015-10-26  Octavian Purdila  <octavian.purdila@intel.com>

    * cofflink.c: Move weak external relocation code to a new function
    (_bfd_coff_relocate_weak_externals)
    * cofflink.c (_bfd_coff_generic_relocate_section): Check if we
    need to relocate a weak external even if the symbol is undefined
    * cofflink.c (coff_link_add_symbols): Fill in the class and aux
    when adding a weak external over an undefined or weakly defined
    symbol