Relocate aux entries for NT weak externals
Currently the tagndx for NT weak externals are not relocated when
creating a relocatable object which breaks relocations when doing the
final link of the relocatable object.

For the following example:

    a.c:
        void foo(void);

            int main()
            {
                    foo();
                    return 0;
            }

    b.c:
            #include <stdio.h>

            void __attribute__((weak)) foo(void)
            {
                    printf("%s weak\n", __func__);
            }

$ i686-w64-mingw32-ld -r -o r.o a.o b.o

This is how the symbol table looks for b.o:

$ i686-w64-mingw32-objdump -x b.o | grep -A 1 foo
[ 23](sec  1)(fl 0x00)(ty   0)(scl   2) (nx 0) 0x00000000 .weak._foo.
[ 24](sec  0)(fl 0x00)(ty  20)(scl 105) (nx 1) 0x00000000 _foo

And this is how the symbol table table looks for r.o:

$ i686-w64-mingw32-objdump -x r.o | grep -A 1 foo
[ 45](sec  0)(fl 0x00)(ty  20)(scl 105) (nx 1) 0x00000000 _foo
AUX lnno 1 size 0x0 tagndx 23
...
[ 48](sec  1)(fl 0x00)(ty   0)(scl   2) (nx 0) 0x00000000 .weak._foo.

As you can see the AUX entry for _foo is wrong in r.o, it should be 48
but it is 23.

The patch uses the sym_indices array to relocate the AUX entries for
weak externals in _bfd_coff_link_input_bfd.

In order to do so we need the .weak. counterpart symbol to be loaded
before the weak external symbol. This is prevented by by the ISFCN()
check _bfd_coff_link_input_bfd and is caused by the fact that the AUX
entry type is 0. To fix this issue we patch gas to propagate the type
of external weak symbols to its .weak. counterpart symbol.

Note that we perform the AUX tag relocations only int the case of
relocatable output because weak external relocation handling
(_bfd_coff_generic_relocate_section) uses the tagndx relative to the
input object.

bfd/

2015-10-26  Octavian Purdila  <octavian.purdila@intel.com>

    * cofflink.c (_bfd_coff_link_input_bfd): relocate AUX entries for
    weak externals when generating relocatable objects

gas/

2015-10-26 Octavian Purdila  <octavian.purdila@intel.com>

    * obj-coff.c (coff_frob_symbol): propagate the type of external
    weak symbol to its .weak. counterpart