Convert NT weak externals to locals
This patch allows converting a weak external symbol to a local symbol
by setting up its section and value to the ones of the alternate
symbol. This is useful when we want to make only a few selected
symbols visible (e.g. objcopy -G).

The conversion is triggered when an NT weak external symbol has been
marked with BSF_LOCAL.

bfd/

2015-10-22  Octavian Purdila <octavian.purdila@intel.com>

    * coffgen.c: add coff_nt_weak_to_local() that implements weak
      external to local symbol conversion
    * coffcode.h (coff_write_object_contents): call
      coff_nt_weak_to_local()
    * libcoff.h: add coff_nt_weak_to_local() declaration

binutils/

2015-10-26  Octavian Purdila <octavian.purdila@intel.com>

    * objcopy.c (filter_symbols): allow converting undefined weak
    symbols to local symbols for PE objects (as PE weak externals are
    implemented as undefined weak symbols)